1. Fixed BigBuffer deletes by adding square brackets. 2. Fixed wxSnprintf usage by adding constants that indicate the size of wxChar buffers
const size_t bufSize = N;
wxChar buf[bufSize];
wxSnprintf(buf, bufSize, ...

instead of using sizeof

wxChar buf[N];
wxSnprintf(buf, sizeof(buf), ...

to prevent a write past the end of the wxChar buffer.  The wxSnprintf
function writes a null character at the end of the buffer to guarantee
data that is too long is terminated.  Since wxChar represents Unicode
strings that are 2 bytes and sizeof(buf) returns the number of characters
not elements in the wxChar buffer, the null was placed past the end of
the array.